SEQUOIADBMAINSTREAM-3749 【Story】SequoiaDB兼容Amazon S3接口 -支持分隔符可配置 代码提交 包括配置，查询分隔符 PUT Object、Delete Object支持按照分隔符生成、删除目录 List Objects，List Versions支持按照分隔符走目录查询，并重构了按照表扫描查询 修改了区域中元数据集合空间+集合的创建方式，创建区域时就创建元数据集合空间，最新元数据集合，历史元数据集合，目录集合，而不是等到上传对象时才创建元数据集合，对象数据集合还是动态创建
git-svn-id: http://192.168.20.11/sequoiadb/trunk@40202 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/sequoias3/src/main/java/com/sequoias3/dao/UserDao.java
@@ -16,7 +16,7 @@ public interface UserDao {
 
     User getUserByAccessKeyID(String accessKeyID) throws S3ServerException;
 
-    Owner getOwnerByUserID(int userId) throws S3ServerException;
+    Owner getOwnerByUserID(long userId) throws S3ServerException;
 
-    int getMaxID() throws S3ServerException;
+    long getMaxID() throws S3ServerException;
 }
