+++ SequoiaDB/engine/tools/sequoias3/src/main/java/com/sequoias3/service/BucketVersioningService.java
@@ -1,11 +1,10 @@
 package com.sequoias3.service;
 
-import com.sequoias3.model.VersioningConfiguration;
 import com.sequoias3.model.VersioningConfigurationNull;
 import com.sequoias3.exception.S3ServerException;
 
 public interface BucketVersioningService  {
-    void putBucketVersioning(int ownerID, String bucketName, String status) throws S3ServerException;
+    void putBucketVersioning(long ownerID, String bucketName, String status) throws S3ServerException;
 
-    VersioningConfigurationNull getBucketVersioning(int ownerID, String bucketName) throws S3ServerException;
+    VersioningConfigurationNull getBucketVersioning(long ownerID, String bucketName) throws S3ServerException;
 }
